% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption.R
\name{build_caption}
\alias{build_caption}
\title{Construire une légende formatée}
\usage{
build_caption(..., nCharByRow = 70)
}
\arguments{
\item{...}{Liste de paires nom-valeur supplémentaires à inclure dans la légende.}

\item{nCharByRow}{Nombre maximal de caractères par ligne pour le formatage (par défaut 70).}
}
\value{
Une chaîne de caractères formatée en HTML représentant la légende.
}
\description{
Cette fonction génère une légende formatée en HTML à partir de divers champs d'entrée.
Elle divise les chaînes de caractères en segments de longueur spécifiée pour un affichage
optimal.
}
\examples{
# Exemple d'utilisation de la fonction buildCaption
\dontrun{
library(ggplot2)
myCaption <- build_caption(Source="SSER",
                           Champ="France métropolitaine",
                           Lecture="Indication pour aider à la lecture du graphique. Ce texte s'il est trés long sera tronqué aprés nCharByRow charactères pour aller sur une 2eme ligne")
ggplot(mpg, aes(x = class, fill = drv)) +
  geom_bar() +
  scale_fill_manual(values = colors_sser("G1","P1","E1")) +
  theme_sser()+
  labs(caption = myCaption)
}


}
