% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs.R
\name{colors_sser}
\alias{colors_sser}
\title{Récupérer ou afficher les couleurs de la charte graphique du SSER.}
\usage{
colors_sser(...)
}
\arguments{
\item{...}{Un vecteur de charactère contenant le nom  des couleurs à
retourner ou une liste enumérée des couleurs à renvoyer (voir exemples).
Si l'input est un vecteur \strong{nommé}, alors le nom est conservé dans le vecteur retourné.}
}
\value{
Un vecteur de couleurs. Le vecteur retourné conserve les noms si l'input est nommé.
}
\description{
La fonction \code{colors_sser} permet de récupérer les codes couleurs de la charte graphique du SSER.
Si aucun argument n'est renseigné, la fonction affiche la grille complete des couleurs disponibles.
}
\examples{
library(ggplot2)
library(chartegraphique.sser)
# Afficher toutes les couleurs de la palette SSER
colors_sser()

# Récupérer les couleurs G1, E1 et P4 :
colors_sser("G1", "E1", "P4")
colors_sser(c("G1", "E1", "P4"))

# Utilisation d'un vecteur nommé :
mes_couleurs_par_annees <- c("annee_2022" = "G1", "annee_2023" = "P1")
colors_sser(mes_couleurs_par_annees)

# appliquer une couleur sur un graphique :
ggplot(mpg, aes(x = class)) +
  geom_bar(fill=colors_sser("P1"))

}
