% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurs.R
\name{display_colors}
\alias{display_colors}
\title{Afficher une grille de couleurs}
\usage{
display_colors(colorVector)
}
\arguments{
\item{colorVector}{Un vecteur nommé de couleurs. Les noms des éléments du vecteur sont utilisés comme étiquettes dans la grille.}
}
\value{
Un graphique ggplot affichant la grille de couleurs.
}
\description{
La fonction \code{display_colors} génère et affiche une grille de couleurs à partir d'un vecteur de couleurs nommées.
Chaque couleur est représentée par un carré dans une grille, avec le nom de la couleur affiché en dessous.
}
\examples{
# Exemple d'utilisation de la fonction display_colors
library(chartegraphique.sser)
couleurs <- c("rouge" = "#FF0000", "vert" = "#00FF00", "bleu" = "#0000FF")
chartegraphique.sser:::display_colors(couleurs)

}
