% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_tableau.R
\name{generer_tableau_insee}
\alias{generer_tableau_insee}
\title{generer_tableau_insee}
\usage{
generer_tableau_insee(
  n_lignes,
  format_colonnes = 1,
  nom_fichier = "tableau_insee.png"
)
}
\description{
genere une image d'un tableau de n lignes, sur 1 ou 2 colonnes
}
