% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inforapide.R
\name{inforap_html}
\alias{inforap_html}
\title{Create a template inforapide in HTML}
\usage{
inforap_html(..., dev = FALSE)
}
\value{
Un objet de format de sortie R Markdown à transmettre
\code{rmarkdown::\link{render}()}.
}
\description{
Create a template inforapide in HTML
}
