% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_typst_extension.R
\name{install_insee_analyses_typst}
\alias{install_insee_analyses_typst}
\title{install_insee_analyses_typst}
\usage{
install_insee_analyses_typst()
}
\description{
Prepare un fichier insee analyses pdf à partir d'un fichier .qmd
}
