% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_sser.R
\name{palettes_sser}
\alias{palettes_sser}
\title{Affichage et accès aux palettes de couleurs sser}
\usage{
palettes_sser(palette = NULL)
}
\arguments{
\item{palette}{Une chaîne de caractères indiquant le nom de la palette souhaitée.
Si \code{NULL} (par défaut), la fonction affiche un graphique de toutes les palettes disponibles.
Les palettes disponibles sont :
\itemize{
\item "IRJ" (4 couleurs)
\item "bleu" (4 couleurs)
\item "vert" (4 couleurs)
\item "orange" (4 couleurs)
\item "prune" (4 couleurs)
\item "IRJ_2_col" (2 couleurs)
\item "IRJ_3_col" (3 couleurs)
\item "IRJ_4_col" (4 couleurs)
\item "IRJ_5_col" (5 couleurs)
\item "IRJ_6_col" (6 couleurs)
}}
}
\value{
Si \code{palette} est \code{NULL}, la fonction affiche un graphique des palettes et retourne \code{NULL} invisiblement.
Sinon, elle retourne un vecteur de couleurs correspondant à la palette demandée.
}
\description{
Cette fonction permet d'afficher toutes les palettes de couleurs disponibles
ou de retourner une palette spécifique si son nom est fourni.
}
\details{
Les palettes disponibles incluent :
\describe{
\item{IRJ}{Palette principale à 4 couleurs (G1, G2, G3, G4)}
\item{bleu}{Palette bleue à 4 couleurs (F1, F2, F3, F4)}
\item{vert}{Palette verte à 4 couleurs (E1, E2, E3, E4)}
\item{orange}{Palette orange à 4 couleurs (O1, O2, O3, O4)}
\item{prune}{Palette prune à 4 couleurs (P1, P2, P3, P4)}
\item{IRJ_2_col à IRJ_6_col}{Palettes IRJ avec 2 à 6 couleurs}
}
}
\examples{
library(ggplot2)
library(chartegraphique.sser)
# Afficher toutes les palettes
palettes_sser()

# Récupérer une palette spécifique
couleurs <- palettes_sser("IRJ")
couleurs <- palettes_sser("IRJ_6_col")

# Appliquer une palette :
ggplot(mpg, aes(x = class, fill = drv)) +
  geom_bar() +
  scale_fill_manual(values = palettes_sser("IRJ"))

}
