% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_ggplot.R
\name{theme_sser}
\alias{theme_sser}
\alias{theme_sser_irj}
\alias{theme_sser_circle}
\title{Thème ggplot2 personnalisé pour les graphiques SSER}
\usage{
theme_sser(
  base_size = 14,
  plot_title = TRUE,
  plot_title_size = base_size + 1,
  caption_size = base_size,
  axis_title = FALSE,
  axis_title_size = base_size,
  axis_title_just = "mc",
  axis_text = "xy",
  axis_text_size = base_size,
  axis_line = FALSE,
  axis_ticks = FALSE,
  grid = "XY",
  plot_legend = TRUE,
  legend_title = FALSE,
  legend_size = base_size,
  legend_position = "bottom"
)

theme_sser_irj(
  base_size = 14,
  plot_title = TRUE,
  plot_title_size = base_size + 1,
  caption_size = base_size,
  axis_title = FALSE,
  axis_title_size = base_size,
  axis_title_just = "mc",
  axis_text = "xy",
  axis_text_size = base_size,
  axis_line = FALSE,
  axis_ticks = FALSE,
  grid = "XY",
  plot_legend = TRUE,
  legend_title = FALSE,
  legend_size = base_size,
  legend_position = "bottom"
)

theme_sser_circle(...)
}
\arguments{
\item{base_size}{Taille de base de la police (en points). Par défaut : 14.}

\item{plot_title}{Si \code{TRUE}, affiche le titre du graphique. Par défaut : \code{TRUE}.}

\item{plot_title_size}{Taille du titre du graphique. Par défaut : \code{base_size + 1}.}

\item{caption_size}{Taille de la légende du graphique. Par défaut : \code{base_size}.}

\item{axis_title}{\code{TRUE} \code{FALSE} ou une combinaison de 'x' 'y'
pour afficher le titre des axes x et/ou y . Par défaut : \code{FALSE}.}

\item{axis_title_size}{Taille des titres des axes. Par défaut : \code{base_size}.}

\item{axis_title_just}{Justification des titres des axes.
Doit être une chaîne de 1 ou 2 caractères parmi \code{c("b", "l", "m", "c", "r", "t")}.
Par exemple, \code{"mc"} pour centré horizontalement et verticalement. Par défaut : \code{"mc"}.}

\item{axis_text}{\code{TRUE} \code{FALSE} ou une combinaison de 'x' 'y'
pour afficher le texte des axes x et/ou y . Par défaut : \code{"xy"}.}

\item{axis_text_size}{Taille du texte des axes. Par défaut : \code{base_size}.}

\item{axis_line}{\code{TRUE} \code{FALSE} ou une combinaison de 'x' 'y'
pour afficher la ligne de l'axe x et/ou y . Par défaut : \code{FALSE}.}

\item{axis_ticks}{Si \code{TRUE}, affiche les graduations des axes. Par défaut : \code{FALSE}.}

\item{grid}{Grilles du graphique (une combinaison de \code{X}, \code{x}, \code{Y}, \code{y} ou \code{FALSE}). Par défaut : \code{"XY"}.}

\item{plot_legend}{Si \code{TRUE}, affiche la légende. Par défaut : \code{TRUE}.}

\item{legend_title}{Si \code{TRUE}, affiche le titre de la légende. Par défaut : \code{FALSE}.}

\item{legend_size}{Taille du texte de la légende. Par défaut : \code{base_size}.}

\item{legend_position}{Position de la légende :
\code{"bottom"}, \code{"left"}, \code{"top"}, \code{"right"}, ou \code{"none"}. Par défaut : \code{"bottom"}.}
}
\description{
Ces fonctions contiennent les paramètres de \code{theme} à appliquer aux graphiques
afin d'être conforme à la charte graphique du  SSER (couleurs de fond, grilles, polices, etc.)
}
\details{
Ce thème respecte strictement la charte graphique SSER :
\itemize{
\item \strong{theme_sser()} : Version de base avec un fond transparent.
\item \strong{theme_sser_irj()} : Version pour les IRJ avec un fond coloré gris clair
\item \strong{theme_sser_circle()} : Version adaptée aux graphiques circulaires (camemberts).
}
}
\examples{
\dontrun{
library(ggplot2)

# Exemple de base avec theme_sser()
ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
  geom_point() +
  labs(title = "Relation longueur/largeur des sépales") +
  theme_sser()

# Exemple avec theme_sser_irj() (fond coloré pour IRJ)
ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
  geom_point() +
  labs(title = "Relation longueur/largeur des sépales") +
  theme_sser_irj()

# Exemple pour un graphique circulaire (camembert) avec theme_sser_circle()
ggplot(data = iris, aes(x = 1, fill = Species)) +
 geom_bar() +
 coord_polar("y") +
 theme_sser_circle()
}

}
\seealso{
\code{\link[ggplot2]{theme}} pour la fonction ggplot2 sous-jacente,
}
